% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LaTeXDatax.r
\name{datax}
\alias{datax}
\title{A function to store a number of data to file}
\usage{
datax(filename, ...)
}
\arguments{
\item{...}{tag-value pairs. The value can be a string, a number or
a list of (value, unit) or (value, unit, format)}

\item{file}{file handle}
}
\value{
nothing
}
\description{
A function to store a number of data to file
}
\examples{
datax(
    "data.tex",
    a = 3, b = "string", c = list(612.2, "nm"),
    d = list(3.14159265358979323846264338327950288419, "", "\%.12f")
)

}
