% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LaTeXDatax.r
\name{print_datum}
\alias{print_datum}
\title{A function to print a single datum to file}
\usage{
print_datum(file, tag, value, unit, format = "\%.4g")
}
\arguments{
\item{file}{File handle of the file to print to}

\item{tag}{a tag to refer to in your document}

\item{value}{the value to connect to tag}

\item{unit}{the unit of value or an empty string}

\item{format}{numeric C-style format string}
}
\value{
nothing
}
\description{
A function to print a single datum to file
}
